% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_tidy.R
\name{augment.arima_tidy}
\alias{augment.arima_tidy}
\title{Augment data with information from an arima_tidy object}
\usage{
\method{augment}{arima_tidy}(fit, newdata, ...)
}
\arguments{
\item{newdata}{The data used to originally fit the model as well as any additional data to be
used to generate predictions. The rows are assumed to be ordered from earliest to latest. If
the originally model was fit on \code{n} data points, the first \code{n} rows of \code{newdata} are assumed to
be the data used to fit the model, and any additional rows are used to generate predictions
using \code{\link{predict.arima_tidy}}.}

\item{...}{Additional arguments. Not used.}
}
\value{
\code{newdata} with additional columns \code{.fitted} and \code{.resid} containing the predicted values
and the residuals.
}
\description{
A method for the \code{\link[generics]{augment}} generic for \code{\link{arima_tidy}} objects.
Takes a model object and a dataset, and adds information about fitted values in the \code{.fitted}
column and model residuals in the \code{.resid} column.
}
\examples{
data(presidents_df)
# Fit on the first 110 rows, predict the last 10
fit <- arima_tidy(approval_rating ~ 1,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
# Augment takes the whole dataset (both training and prediction data)
augment(fit, presidents_df)

fit <- arima_tidy(approval_rating ~ election_year,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
augment(fit, presidents_df)
}
