% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_helpers.R
\name{calculate_alarm_type}
\alias{calculate_alarm_type}
\title{Identify the category of an alarm function}
\usage{
calculate_alarm_type(alarm_function_name)
}
\arguments{
\item{alarm_function_name}{The name of the alarm function in question, as a character string.}
}
\value{
Either "scan" or "parallel".
}
\description{
`statsurv` uses two different forms of alarm functions: scan-type alarms, which
  calcualte an alarm statistic over a large number of spatial zones, and parallel-type alarms,
  which calculate an alarm statistic over individual spatial locations. These are identifiable by
  whether the name of the alarm function starts with "scan_" or "parallel_".
}
\examples{
calculate_alarm_type("scan_eb_poisson")
calculate_alarm_type("parallel_cusum_poisson")
}
\keyword{internal}
