% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_county_fips_2019}
\alias{co_county_fips_2019}
\title{Geographical boundaries for counties in Colorado}
\format{
A SpatialPolygonsDataFrame with 64 rows and 1 variables:
\describe{
  \item{fips_str}{The 5-digit code used to identify the county. A full list of IDs is available
  from the
  \href{https://www.census.gov/geographies/reference-files/2018/demo/popest/2018-fips.html}{US
  Census Bureau} ("08001" - "08125")}
 }
}
\source{
\url{https://data-cdphe.opendata.arcgis.com/datasets/colorado-county-boundaries}
}
\usage{
co_county_fips_2019
}
\description{
A SpatialPolygonsDataFrame containing the geographical boundaries for all counties
  in Colorado, as of 2019.
}
\note{
To save memory, county boundaries were simplified using the \code{\link[sf]{st_simplify}}
   command and therfore do not exactly match the source shapefiles.
}
\keyword{datasets}
