% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_hsr_2019}
\alias{co_hsr_2019}
\title{Geographic boundaries for Colorado health statistics regions}
\format{
A SpatialPolygonsDataFrame with 21 rows and 1 variables:
\describe{
  \item{hsr}{A string labeling the region, of the form "hsr_XX" ("hsr_01" - "hsr_21")}
}
}
\source{
\url{https://data-cdphe.opendata.arcgis.com/datasets/cdphe-colorado-health-statistics-regions}
}
\usage{
co_hsr_2019
}
\description{
A SpatialPolygonsDataFrame containing the geographic boundaries for all health
  statistics regions in Colorado, as of 2019. The health statistics regions are aggregations of
  demographically similar counties and divide the state into 21 non-overlapping regions.
}
\note{
To save memory, region boundaries were simplified using the \code{\link[sf]{st_simplify}}
   command and therfore do not exactly match the source shapefiles.
}
\keyword{datasets}
