% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_state_boundary}
\alias{co_state_boundary}
\title{Geographic boundary of the state of Colorado}
\format{
A SpatialPolygonsDataFrame with 1 row and 1 variables:
\describe{
  \item{name}{The name of the state ("colorado")}
}
}
\source{
\url{https://data-cdphe.opendata.arcgis.com/datasets/colorado-state-boundary}
}
\usage{
co_state_boundary
}
\description{
A SpatialPolygonsDataFrame containing the geographic boundaries of the state of
  Colorado.
}
\keyword{datasets}
