% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Merge entries into a list of length 1}
\usage{
collapse(vec, combine_dfs = TRUE, unique_only = FALSE)
}
\arguments{
\item{vec}{An atomic vector or a list.}

\item{combine_dfs}{Logical. Whether columns containtaining data frames should be combined into a
single longer data frame (\code{combine_dfs = TRUE}) or left as a list of data frames (\code{combine_dfs = FALSE})}

\item{unique_only}{Logical. Whether the entries in a list should contain the values from all
rows, including repeats (\code{unique_only = FALSE}) or if only unique values should be reported
(\code{unique_only = TRUE}).}
}
\value{
A list of length one. If `combine_dfs = FALSE, unique_only = FALSE`, then `collapse()`
  returns a `list(vec)`. If `combine_dfs = TRUE`, and `vec` is a list of data frames, then
  `collapse()` returns `list(bind_rows(vec))`. If `unique_only = TRUE`, then duplicate entries of
  `vec` are removed.
}
\description{
Take a vector or list and combine the entries into a list of length 1.
}
\details{
`collapse()` is a helper function used by \code{\link{collapse_all}}, but can also be used in
conjunction with \code{\link[dplyr]{summarize}} to reduce the number of rows in a data frame
without losing information.
}
\examples{
vec <- c(1, 2, 3, 3, 4)
collapse(vec)

collapse(vec, unique_only = TRUE)

df1 <- data.frame(x = c(1, 2, 3), y = c("A", "B", "C"))
df2 <- data.farme(x = c(4, 5), y= c("D", "E"))
collapse(list(df1, df2), combine_dfs = TRUE)

# collapse can be used together with summarize:
mtcars_nested <- mtcars \%>\%
  group_by(cyl) \%>\%
  nest()

# Generate two models for mpg with different predicting variables
model_fits <- tidyr::crossing(mtcars_nested, model_var = c("wt", "hp")) \%>\%
  rowwise() \%>\%
  mutate(model = list(lm(paste0("mpg ~ ", model_var), data = data)),
         predictions = list(broom::augment(model)))

# Combine the results from each model separately
model_fits \%>\%
  group_by(model_var) \%>\%
  summarize(models = collapse(model),
            predictions = collapse(predictions, combine_dfs = TRUE))
}
