% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_yhat.R
\name{collapse_yhat}
\alias{collapse_yhat}
\title{Average sampled model predictions}
\usage{
collapse_yhat(space_coord, time_coord, yhat)
}
\arguments{
\item{space_coord}{A gridcoord object (\code{\link[gridcoord]{gc_gridcoord}}) describing the
spatial area that is covered by \code{yhat}.}

\item{time_coord}{A gridcoord object (\code{\link[gridcoord]{gc_gridcoord}}) describing the time
span that is covered by \code{yhat}.}

\item{yhat}{A data.frame created by \code{\link{sample_yhat}} whose first two columns correspond
to \code{space_coord} and \code{time_coord}, and whose remaining columns are random samples of the model
predictions.}
}
\value{
A data.frame whose first two columns correspond to \code{space_coord} and \code{time_coord}, and
whose third column is the mean of all the samples in \code{yhat}.
}
\description{
Average model predictions generated by \code{\link{sample_yhat}}.
}
\examples{
x <- rnorm(20, mean = 0, sd = 5)
y <- 1.3 + 2.5 * x + rnorm(20, sd = 1)
fit <- lm(y ~ x)

new_data <- data.frame(space = c("s1"),
                       time = c("t1", "t2", "t3"),
                       x = c(-3, 3, 8),
                       stringsAsFactors = FALSE)
space_coord <- data.frame(space = "s1",
                          stringsAsFactors = FALSE)
time_coord <- data.frame(time = c("t1", "t2", "t3"),
                         stringsAsFactors = FALSE)

yhat <- sample_yhat(space_coord,
                    time_coord,
                    fit,
                    new_data,
                    n_samples = 10)

collapse_yhat(space_coord, time_coord, yhat)
}
\seealso{
\code{\link{sample_yhat}}
}
