% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_zones.R
\name{create_zones}
\alias{create_zones}
\title{Create zones from spatial coordinates}
\usage{
create_zones(sf, max_k, min_k = 1)
}
\arguments{
\item{sf}{A simple-features (sf) collection describing the spatial locations to combine into
zones. Each row of \code{sf} should represent a unique spatial location.}

\item{max_k}{The maximum number of locations to be included in any one zone.}

\item{min_k}{Optional. The minimum number of locations to be included in any one zone.}
}
\value{
A list of integer vectors, specifying which locations belong to which zones. The integers
identify locations based on which row in \code{sf} they are associated with.
}
\description{
Scan-type alarm functions work by computing an alarm function over a very large
number of spatial \emph{zones}, where each zone consists of one of more neighboring spatial
locations. Zones can be many sizes, from a single location up to an aggregation of \code{max_k}
locations, and each location can belong to multiple zones. \code{create_zones} aggregates
individual spatial locations into a set of zones using a k-nearest neighbors approach. For each
location, there will be a zone containing the location itself, a zone containing the location
and its nearest neighbor, a zone containing the location and its two nearest-neighbors, up to
the maximum size.
}
\examples{
library("sf")
data("NM_county_sf")
create_zones(NM_county_sf, max_k = 5, min_k = 2)
# 112 zones created from 32 locations
}
