% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denver_censustract_fips_2019}
\alias{denver_censustract_fips_2019}
\title{Geographical boundaries for census tracts in the Denver Metro Area}
\format{
A SpatialPolygonsDataFrame with 534 rows and 1 variables:
\describe{
  \item{fips_str}{The 11-digit code used to identify the county ("08001" - "08125")}
 }
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
denver_censustract_fips_2019
}
\description{
A SpatialPolygonsDataFrame containing the geographical boundaries for all census
  tracts in the Denver Metro Area. The approximate geographic boundary for metro area was defined
  as the beltway formed by highways CO-93, CO-470, E-470, the Northwest Parkway, US-36, and
  CO-170. Any census tract with any portion inside this ring is included in this dataset.
}
\keyword{datasets}
