% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denver_metro_area}
\alias{denver_metro_area}
\title{Approximate boundary of the Denver metro area}
\format{
A SpatialPolygonsDataFrame with 1 rows and 1 variables:
\describe{
  \item{Name}{NA}
 }
}
\usage{
denver_metro_area
}
\description{
A SpatialPolygonsDataFrame describing the approximate geographic boundary for metro
  area was defined as the beltway formed by highways CO-93, CO-470, E-470, the Northwest Parkway,
  US-36, and CO-170.
}
\keyword{datasets}
