% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_helpers.R
\name{ensure_zones}
\alias{ensure_zones}
\title{Ensure that zone information is repesented as a list}
\usage{
ensure_zones(zone_info)
}
\arguments{
\item{zone_info}{Either:
\itemize{
\item A list where each entry is a zone. Each entry is a vector containing the indices of
the spatial locations in that zone.
\item A matrix, where each row is a zone and each column is a location. The entry \verb{[i, j]}
is 1 if location \code{j} is a part of zone \code{i} and 0 otherwise.
}}
}
\value{
A list, where each entry is a zone.
}
\description{
Information about which spatial locations belong to which zone can be specified in
the form of a list, where each entry in the list contains the indexes of the locations in a
particular zone, or in the form of a matrix, where each row represents a zone and each column
represents a location. \code{ensure_zones} accepts either representation as input, and always
returns zones represented as a list.
}
\examples{
library("sf")
zone_list <- space_coord_to_zones(co_county_fips_2019, 5)
zone_matrix <- build_key_matrix(zone_list)

\dontrun{
# Both give the same results
ensure_zones(zone_list)
ensure_zones(zone_matrix)
}
}
