% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_alarm_statistic.R
\name{extract_alarm_statistic}
\alias{extract_alarm_statistic}
\title{Summarize information from multiple alarm functions}
\usage{
extract_alarm_statistic(space_coord, time_coord, list_of_alarms, zone = NULL)
}
\arguments{
\item{space_coord}{A gridcoord object
(\code{\link[gridcoord]{gc_gridcoord}}) describing the spatial area that is
covered by \code{list_of_alarms}}

\item{time_coord}{A gridcoord object (\code{\link[gridcoord]{gc_gridcoord}})
describing the time points that are covered by \code{list_of_alarms}. This coordinate must be
an ordered dataframe, with the first entries corresponding to the earliest time periods and
the last entries corresponding to the most recent.}

\item{list_of_alarms}{A list of scan-type alarm function results, as generated by
\code{\link{loop_alarm_function}}. Each alarm function result must contain the components
\code{observe} and \code{replicates}, and both of these components must contain the column
\code{action_level}. This can be ensured by generating \code{list_of_alarms} with
\code{\link{loop_alarm_function}} or \code{\link{standardized_alarm_functions}}.}

\item{zone}{The index of the zone to report the alarm statistic for. If NULL (default), reports
the zone with the highest value of the alarm statistic at each time point.}
}
\value{
A data.frame containing one row per time point. In addition to the columns contained in
the alarm function results and a column corresponding to the time coordinate, the data.frame
will contains the following columns:
\describe{
\item{upper_control_limit}{The 97.5 percentile of the scan statistic calcluated by Monte Carlo
simulation. An \code{action_level} above this value is evidence that the observed behavior does not
match what would be expected based on the model. If the scan statistic was calculated with the
parameter \code{n_mcsim = 0}, then this column with be \code{NA}.}
\item{alert_start_label}{The label of the time coordinate when the potential outbreak began.}
}
If \code{time_coord} is a span-type gridcoord, then data.frame will contain two additional columns:
\describe{
\item{surveillance_date}{The start date of the span when the surveillance was performed.}
\item{alert_start_date}{The start date of the time span when the potential outbreak began.}
}
}
\description{
\code{extract_alarm_statistic} combines information from a list of alarm statistics
calculated at different times to provide an overview about the alarm statistic in a particular
zone over time or the maximum value of the alarm statistic over time.
}
\examples{
# Using pre-computer alarm statistics for New Mexico
nm_county_fips_2010 <- statsurv::nm_county_coord
nm_scan_alarms <- statsurv::nm_scan_alarms

year_coord <- generate_date_range(lubridate::ymd("1973-01-01"),
                                  lubridate::ymd("1991-01-01"),
                                  time_division = "year")
extract_alarm_statistic(nm_county_fips_2010, year_coord, nm_scan_alarms)
extract_alarm_statistic(nm_county_fips_2010, year_coord, nm_scan_alarms, zone = 4)
}
