% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clusters.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract clusters identified by scanning alarm functions}
\usage{
extract_clusters(
  scan_result,
  zones,
  use_column = c("action_level", "mc_pvalue", "score", "log_posterior"),
  cutoff = 0.1,
  max_allowed = length(zones),
  allow_overlap = TRUE
)
}
\arguments{
\item{scan_result}{The result from a scan-type alarm function
(\code{\link{standardized_alarm_functions}}).}

\item{zones}{A list of integer vectors. Each vector corresponds to a single zone; its elements
are the numbers of the locations in that zone. Must be generated as part of calculating the
scan statistic.}

\item{use_column}{How should clusters be identified? Allowed options are "action_level",
"pvalue_mc", "score", and "log_posterior". "action_level" is valid for all scanstatistic
objects returned by \code{\link{standardized_alarm_functions}}, "pvalue_mc" and "score" are
valid for all scanstatistics except for \code{\link[scanstatistics]{scan_bayes_negbin}}, and
"log_posterior" is valid only for \code{\link[scanstatistics]{scan_bayes_negbin}}.}

\item{cutoff}{The minimun or maximum value of \code{use_column} to be defined as a cluster. Values
must be below \code{cutoff} if \code{use_column = "pvalue_mc"}; otherwise values must be above \code{cutoff}.}

\item{max_allowed}{The maximum number of clusters to be reported.}

\item{allow_overlap}{Boolean. If FALSE, clusters are only reported if they have no spatial
regions in common with a more intense cluster.}
}
\value{
A \code{data.frame} with at most \code{max_allowed} rows, and including columns \code{zone}, \code{duration},
and whichever column was specified by \code{use_column}.
}
\description{
Extract all space-time clusters, based on values calculated by a scan statistic.
This can either be an absolute measure of intensity, such as the score calculate by
\code{\link[scanstatistics]{scan_eb_poisson}} or a relative measure, such as the Monte-Carlo
p-value.
}
\details{
A cluster is defined as a spatio-temporal zone with a calculated value above or below a specified
cutoff score. One cluster is always returned, even if its value does not meet the cutoff
requirements, and at most \code{max_allowed} clusters are reported. A cluster is never reported if it
covers exactly the same spatial zone but has a different duration as another more intense
cluster. If \code{allow_overlap} is FALSE, clusters are only reported if they have no spatial regions
in common with a more intense cluster.
}
\examples{
# Using pre-computer alarm statistics for New Mexico
alarm <- statsurv::nm_scan_alarms[["1989"]]

# By default, use the scan statistics score as the ranking mechanism:
extract_clusters(alarm,
                 alarm$zone_info)

# Use a cutoff of a p-value of < 0.05:
extract_clusters(alarm,
                 alarm$zone_info,
                 use_column = "mc_pvalue")
}
