% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_yhat.R
\name{extract_yhat.forecast_ARIMA}
\alias{extract_yhat.forecast_ARIMA}
\title{Extract fitted values from a forecast_ARIMA model object}
\usage{
\method{extract_yhat}{forecast_ARIMA}(fit, data, xreg)
}
\arguments{
\item{fit}{A model fit object, such as that generated by \code{lm} or \code{glm}.}

\item{data}{A data.frame containig the data that should be used to predict the fitted values.
Must have all of the covariates used by the model fit object to calculate fitted values, as
well as columns corresponding to \code{space_coord} and \code{time_coord}}

\item{xreg}{A matrix of covariate information used to calculate ARIMA values. Must cover the
exact same temporal and spatial locations as \code{data}}

\item{n_ahead}{How many data points should be predicted?}

\item{...}{Other arguments}
}
\value{
A data.frame whose first two columns correspond to \code{space_coord} and \code{time_coord}, and
whose remaining columns are the point estimates or random samples of the fitted values.
}
\description{
Extract fitted values from a forecast_ARIMA model object
}
\details{
ARIMA models currently only have limited support in the statsurv package. The syntax for
fitting and working with ARIMA models is significantly different than that of \code{lm} or \code{glm}
models. Some of the key differences are:
\enumerate{
\item extract_yhat.forecast_ARIMA only works on models containing a single time series. In
other words, the space coordinate must only have a single location.
\item forecast_ARIMA models do not use the \code{data} parameter, instead any covariates must be
included in the matrix \code{xreg}
\item extract_yhat.forecast_ARIMA needs to know how many time-steps ahead should be predicted,
as specified by the \code{n_ahead} parameter.
}
}
\examples{
\dontrun{
library("scanstatistics")
library("forecast")
nm_county_coord <- statsurv::nm_county_coord
data(NM_popcas)
year_coord <- generate_date_range(lubridate::ymd("1973-01-01"),
                                  lubridate::ymd("1991-01-01"),
                                  time_division = "year")
year_coord$year <- year_coord$date_label
year_coord <- gridcoord::gc_gridcoord(year_coord, "year")

# Fit the data for Santa Fe county via an ARIMA model,
# not including the last 2 data points:
santa_fe <- NM_popcas \%>\%
  dplyr::filter(county == "santafe")
sf_coord <- nm_county_coord \%>\%
  dplyr::filter(county == "santafe")
n <- seq(1, nrow(santa_fe) - 2)
xreg = as.matrix(santa_fe[n, "population", drop = FALSE])
fit_Arima <- Arima(santa_fe$count[n],
                   order = c(2, 1, 0),
                   xreg = xreg[n, , drop = FALSE])

# Then use extract_yhat to generate predictions:
# Because we did not include the last 2 points in fitting the model,
# we set n_ahead = 2
extract_yhat(sf_coord, year_coord, fit_Arima, santa_fe,
             xreg, n_ahead = 2)
}
}
\seealso{
\code{\link{extract_yhat}}, \code{\link[forecast]{forecast}}
}
