% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_yhat.R
\name{extract_yhat.inla}
\alias{extract_yhat.inla}
\title{#' @export
extract_yhat.lm <- function(fit, data, ...) {
  data$.fitted <- stats::predict(fit, newdata = data, type = "response")
  return(data)
}}
\usage{
\method{extract_yhat}{inla}(fit, data)
}
\description{
#' @export
extract_yhat.lm <- function(fit, data, ...) {
  data$.fitted <- stats::predict(fit, newdata = data, type = "response")
  return(data)
}
}
