% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coordinates.R
\name{generate_study_area}
\alias{generate_study_area}
\title{Generate a set of spatial coordinates}
\usage{
generate_study_area(
  surveillance_boundary = NULL,
  space_division = "co_censustract_fips_2019",
  inside_fraction_cutoff = 0
)
}
\arguments{
\item{surveillance_boundary}{A SpatialPolygon* or sf object that defines the boundary of the
study area. If NULL, the entirity of \code{space_division} will be returned.}

\item{space_division}{What spatial divisions should be used. Allowed options are
'denver_censustract_fips_2019', 'co_county_fips_2019', 'co_hsr_2019', 'co_state_boundary', or a
spatial polygons data frame or sf object. The 4 non-filename options instruct the function to
use sptial divisions that are included with the `statsurv`` package, and were processed on 19
May 2020. They correspond to all of the census tracts around Denver and the counties, health
statistics regions, and states within Colorado.}

\item{inside_fraction_cutoff}{What fraction of the unit must be inside the surveillance_boundary
for the unit to be included? Default is 0 (any unit that touches the boundary will be
included).}
}
\value{
A spatial polygons data frame that is a valid
\code{\link[gridcoord:gc_gridcoord]{gridcoord}} object. A spdf object is returned, regardless
of the format of \code{surveillance_boundary} or \code{space_division}.
}
\description{
Generate a set of spatial coordinates, given a collection of polygons and optionally
a spatial boundary.
}
\examples{
surveillance_area <- statsurv::denver_metro_area

# Returns all counties that have at least 10\% of their area within
# the greater Denver Metro area
generate_study_area(surveillance_area,
                    "co_county_fips_2019",
                    inside_fraction_cutoff = 0.1)

# Returns all of the Colorado Health Statistics Regions
generate_study_area(space_division = "co_hsr_2019")
}
