% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inla.R
\name{glance_inla}
\alias{glance_inla}
\title{Summarize an INLA model}
\usage{
glance_inla(fit)
}
\arguments{
\item{fit}{The model fit object generated by INLA.}
}
\value{
A 1-row data frame containing information on goodness of fit, p-values, or model
convergence for the model object. Contains information on DIC, WAIC, the log likelihood, CPO,
and the effective number of parameters.
}
\description{
\code{\link[broom]{glance}} summarizes a model into a data frame. \code{glance_inla} adapts this to
INLA models.
}
\examples{
x <- rnorm(20, mean = 0, sd = 3)
y <- 1.3 + 2.5 * x + rnorm(20, sd = 1)
df <- data.frame(x = x, y = y)

\dontrun{
fit_inla <- INLA::inla(y ~ x, data = df)
glance_inla(fit_inla)

data(NM_data)
fit_inla <- INLA::inla(count ~ population + f(county, model = "iid"),
                       data = NM_data,
                       control.compute = list(dic = TRUE, cpo = TRUE, waic = TRUE))
glance_inla(fit_inla)
}
}
