% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_county_coord}
\alias{nm_county_coord}
\title{Spatial coordinate corresponding to the dataset NM_popcas}
\format{
A Simple Features collection with 32 rows and 2 variables:
\describe{
  \item{county}{The name of the county in all lower case and spaces removed
  ("bernalillo" - "valencia")}
 }
}
\source{
\url{http://rgis.unm.edu/rgis6/}
}
\usage{
nm_county_coord
}
\description{
A simple features collection containing the geographical boundaries for all counties
  in New Mexico that are included in the dataset \code{\link[scanstatistics]{NM_popcas}}
}
\note{
To save memory, county boundaries were simplified using the \code{\link[sf]{st_simplify}}
   command and therfore do not exactly match the source shapefiles.
}
\keyword{datasets}
