% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_county_fips_2010}
\alias{nm_county_fips_2010}
\title{Geographical boundaries for counties in New Mexico}
\format{
A Simple Features collection with 33 rows and 2 variables:
\describe{
  \item{fips_str}{The 5-digit code used to identify the county. A full list of IDs is available
  from the
  \href{https://www.census.gov/geographies/reference-files/2018/demo/popest/2018-fips.html}{US
  Census Bureau} ("35001" - "35061")}
  \item{county_name}{The name of the county, with the first letter capitalized
  ("Bernalillo" - "Valencia")}
 }
}
\source{
\url{http://rgis.unm.edu/rgis6/}
}
\usage{
nm_county_fips_2010
}
\description{
A simple features collection containing the geographical boundaries for all counties
  in New Mexico, as of 2010
}
\note{
To save memory, county boundaries were simplified using the \code{\link[sf]{st_simplify}}
   command and therfore do not exactly match the source shapefiles.
}
\keyword{datasets}
