% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_county_sf}
\alias{NM_county_sf}
\title{Geographical boundaries for counties in New Mexico}
\format{
A Simple Features collection with 32 rows and 2 variables:
\describe{
  \item{id_space}{An integer label associated with each county (1-32)}
  \item{county}{The name of the county, in all lower case
  ("bernalillo" - "valencia")}
 }
}
\source{
\url{http://rgis.unm.edu/rgis6/}
}
\usage{
NM_county_sf
}
\description{
A simple features collection containing the geographical boundaries for all counties
  in New Mexico, as of 2010. To avoid errors when looking at counties over time, Cibola county is
  not included.
}
\note{
To save memory, county boundaries were simplified using the \code{\link[sf]{st_simplify}}
   command and therefore do not exactly match the source shapefiles.
}
\keyword{datasets}
