% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_cusum_alarms}
\alias{nm_cusum_alarms}
\title{CUSUM results for brain cancer in New Mexico, 1980 - 1991}
\format{
A list where each entry is either \code{NA} or the output of
\code{\link{parallel_cusum_poisson}}, applied to cases of brain cancer in New Mexico
at the county level.
}
\usage{
nm_cusum_alarms
}
\description{
A list containing the results of repeated applying the CUSUM alarm statistic to
cases of brain cancer in New Mexico at the county level. Baseline estimates were generated
using a Poisson model, where the number of cases is a function of year and the population of
the county was used at the offset.
}
\keyword{datasets}
