% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NM_data}
\alias{NM_data}
\title{Population and brain cancer cases in New Mexico, 1973 - 1991}
\format{
A data frame with 608 rows and 7 columns:
\describe{
  \item{id_time}{An integer label associated with each year (1-19)}
  \item{id_space}{An integer label associated with each county (1-32)}
  \item{year}{The year the cases were recorded (1973-1991)}
  \item{county}{The name of the county, in all lower cases, with no spaces ("bernalillo" -
  "valencia")}
  \item{population}{The popoulation of the county in that year. Interpolated from censuses in
  1973, 1982, and 1991 (977-490248)}
  \item{count}{The number of brain cancer cases reported in that county and year (0-34)}
  \item{baseline_est}{A simple baseline estimate of the expected number of cases based on the
  population (0.044-22.5)}
 }
}
\source{
\url{https://github.com/BenjaK/scanstatistics}
}
\usage{
NM_data
}
\description{
A data set containing the population count an dnumber of brain cancer cases in each
  county in New Mexico for the years 1973 - 1991. The data has been formatted to be a valid
  \code{\link[=spacetime_data]{spacetime data frame}}
}
\keyword{datasets}
