% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_scan_alarms}
\alias{nm_scan_alarms}
\title{Scanstatistics results for brain cancer in New Mexico, 1980 - 1991}
\format{
A list where each entry is either \code{NA} or the output of
\code{\link[scanstatistics]{scan_eb_poisson}}, applied to cases of brain cancer in New Mexico
at the county level.
}
\usage{
nm_scan_alarms
}
\description{
A list containing the results of repeated applying the Poisson scan statistic to
cases of brain cancer in New Mexico at the county level. Baseline estimates were generated
using a Poisson model, where the number of cases is a function of year and the population of
the county was used at the offset. The maximum zone size is 3 counties.
}
\keyword{datasets}
