% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_data.R
\name{pad_spacetime_data}
\alias{pad_spacetime_data}
\title{Pad a spacetime data frame}
\usage{
pad_spacetime_data(df)
}
\arguments{
\item{df}{A data frame or a data frame extension (e.g., a tibble) that includes
columns \code{id_space} and \code{id_time} that are each integers. Unlike \code{\link{spacetime_data}},
there may be gaps or skips in the entries in these columns.}
}
\value{
A data frame padded with extra rows to be a valid spacetime data frame. After padding,
all  combinations of \code{id_space} and \code{id_time} are associated with at least one row, and there
are no skipped values in those two columns. Added rows contain \code{NA} for all columns other than
\code{id_space} and \code{id_time}.
}
\description{
Pad a spacetime data frame
}
\examples{
df <- data.frame(id_space = c(1, 2),
                 id_time = c(1, 3),
                 y = c(1, 3))
padded_data <- pad_spacetime_data(df)
validate_spacetime_data(padded_data)
}
\seealso{
\code{\link{spacetime_data}} \code{\link{validate_spacetime_data}}
}
