% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cusum.R
\name{parallel_cusum_gaussian}
\alias{parallel_cusum_gaussian}
\title{Calculate the Gaussian CUSUM statistic over multiple spatial regions}
\usage{
parallel_cusum_gaussian(
  wide_cases,
  wide_baseline,
  mean = 0,
  sigma = 1,
  drift = 0.5
)
}
\arguments{
\item{wide_cases}{A matrix of observed counts. Rows indicate time and are ordered from least
recent (row 1) to most recent (row \code{nrow(counts)}). Columns indicate locations, numbered from 1
and up. Can be generated from a data.frame with the \code{\link{pivot_for_scan}} function.}

\item{wide_baseline}{A matrix of the same dimensions as \code{wide_cases}, holding the expected value
of cases at each space-time location.This value is typically estimated from past data using a
GLM or other model.}

\item{mean}{The estimated mean of the residuals (\code{wide_cases - wide_baseline})}

\item{sigma}{The estimated standard deviation of the residuals.}

\item{drift}{The downward drift to use in calculating the CUSUM statistic.  The value of \code{drift}
is typically half of the expected increase in the value of the standardized residuals during an
outbreak.}
}
\value{
A matrix with the same dimensions as \code{wide_cases}, where rows indicate time (oldest to
newst) and columns indicate locations. Element \verb{(i, j) } indicates the value of the CUSUM
statistic at time \code{i} in region \code{j}. The value of the CUSUM statistic at row \code{i} incorporates
informations from time points \code{1} thought \code{i}.
}
\description{
\code{parallel_cusum_gaussian} calculates the Gaussian CUSUM statistic independently over
each spatial region, over all time points. Unlike e.g., \code{\link{scan_cusum_poisson}},
information from multiple spatial regions is never aggregated.
}
\details{
The CUSUM statistic is widely used surveillance technique for purely temporal processes.
It is calculated by taking the cumulative sum of the standard residuals with a downward drift
and a floor of 0. For a Gaussian process, residuals are calculated as \code{resid = wide_cases - wide_baseline}, and standard residuals are calculated as \code{(resid - mean)/sigma}. The CUSUM
statistic is then calculated by summing the standardized residuals with downward drift \code{drift}.

The three parameters used in calculating the standardized residuals and CUSUM statistic,
\code{mean}, \code{sigma}, and \code{drift}, can each be a scalar, a vector with a separate value for each
spatial region, or a matrix with the same dimensions as \code{wide_cases}
}
\examples{
outbreak_sp <- c(1, 2, 4, 5)
outbreak_tm <- c(3, 4)
wide_cases <- matrix(2, nrow = 4, ncol = 9)
wide_cases[outbreak_tm, outbreak_sp] <- 5
wide_cases[c(3, 4), c(8, 9)] <- 2.5
wide_baseline <- matrix(2, nrow = 4, ncol = 9)
scanres <- parallel_cusum_gaussian(wide_cases, wide_baseline,
                                   mean = 1,
                                   sigma = 2,
                                   drift = 0.5)

v_mean <- c(rep(0.0, 8), 2)
m_sigma <- matrix(1.5, nrow = 4, ncol = 9)
m_sigma[c(3, 4), c(4, 5)] <- 0.25
scanres <- parallel_cusum_gaussian(wide_cases, wide_baseline,
                                   mean = v_mean,
                                   sigma = m_sigma,
                                   drift = 0.5)
}
\seealso{
\code{\link{scan_cusum_poisson}}, \code{\link{pivot_for_scan}},
\code{\link{build_key_matrix}}, \code{\link{parallel_cusum_gaussian}}
}
