% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel2.R
\name{parallel_cusum_gaussian2}
\alias{parallel_cusum_gaussian2}
\title{Calculate the Gaussian CUSUM statistic over multiple spatial regions}
\usage{
parallel_cusum_gaussian2(
  spacetime_data,
  outcome_col,
  baseline_col = NULL,
  mean = 0,
  sigma = 1,
  drift = 0.5
)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{outcome_col}{The column in \code{data} containing the outcome of interest, such as observed counts}

\item{baseline_col}{Optional. The column in \code{data} containing baseline estimates of the outcome
of interest. If not present, assumed to be 0 .}

\item{mean}{The estimated mean of the residuals (\code{spacetime_data$outcome_col - spacetime_data$baseline_col}). Can either be a scalar value or a column in \code{spacetime_data}}

\item{sigma}{The estimated standard deviation of the residuals. Can either be a scalar value or a
column in \code{spacetime_data}}

\item{drift}{The downward drift to use in calculating the CUSUM statistic.  The value of \code{drift}
is typically half of the expected increase in the value of the standardized residuals during an
outbreak. Can either be a scalar value or a column in \code{spacetime_data}}
}
\value{
\code{spacetime_data} with the added column \code{.action_level} containing the value of the CUSUM
statistic at that point and time.
}
\description{
\code{parallel_cusum_gaussian} calculates the Gaussian CUSUM statistic independently over each spatial
region, over all time points. Each spatial region is considered independently.
}
\details{
The CUSUM statistic is widely used surveillance technique for purely temporal processes.
It is calculated by taking the cumulative sum of the standard residuals with a downward drift
and a floor of 0. For a Gaussian process, residuals are calculated as \code{resid = outcome - baseline}, and standard residuals are calculated as \code{(resid - mean)/sigma}. The CUSUM
statistic is then calculated by summing the standardized residuals with downward drift \code{drift}.

The three parameters used in calculating the standardized residuals and CUSUM statistic,
\code{mean}, \code{sigma}, and \code{drift}, can each either be a scalar or a column in \code{spacetime_data}.
}
\examples{
data(NM_data)

parallel_cusum_gaussian2(NM_data, count, drift = 1)
parallel_cusum_gaussian2(NM_data, count, baseline_est)
parallel_cusum_gaussian2(NM_data, count,
                         population * mean(count) / mean(population),
                         sigma = 2)
}
\seealso{
parallel_cusum_gaussian

Other parallel2_alarms: 
\code{\link{parallel_cusum_poisson2}()},
\code{\link{parallel_shewhart_gaussian2}()}
}
\concept{parallel2_alarms}
