% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_shewhart.R
\name{parallel_shewhart_gaussian}
\alias{parallel_shewhart_gaussian}
\title{Calculate the Gaussian Shewhart alarm statistic over multiple spatial regions}
\usage{
parallel_shewhart_gaussian(
  wide_cases,
  wide_baseline = NULL,
  mean = 0,
  sigma = 1
)
}
\arguments{
\item{wide_cases}{A matrix of observed counts. Rows indicate time and are ordered from least
recent (row 1) to most recent (row \code{nrow(counts)}). Columns indicate locations, numbered from 1
and up. Can be generated from a data.frame with the \code{\link{pivot_for_scan}} function.}

\item{wide_baseline}{A matrix of the same dimensions as \code{wide_cases}, holding the expected value
of cases at each space-time location.This value is typically estimated from past data using a
GLM or other model.}

\item{mean}{The estimated mean of the residuals (\code{wide_cases - wide_baseline})}

\item{sigma}{The estimated standard deviation of the residuals.}
}
\value{
A matrix with the same dimensions as \code{wide_cases}, where rows indicate time (oldest to
newst) and columns indicate locations. Element \verb{(i, j) } indicates the value of the Shewhart
alarm statistic at time \code{i} in region \code{j}. The value of the Shewhart statistic at row \code{i}
incorporates informations only from time point \code{i}.
}
\description{
\code{parallel_cusum_gaussian} calculates the value of the Shewhart alarm statistic
assuming a Gaussian distribution independently over each spatial region, over all time points.
The Shewhart alarm statistic considers each spatial region and time point independently, and
information from multiple observations is never aggregated.
}
\details{
The Shewhart alarm statistic is equal to the number of standard deviations that a given
observation is above or below the mean, also known as a z-score. It can be directly calculated
as \code{(wide_cases - wide_baseline - mean) / sd}.
}
\examples{
outbreak_sp <- c(1, 2, 4, 5)
outbreak_tm <- c(3, 4)
wide_cases <- matrix(2, nrow = 4, ncol = 9)
wide_cases[outbreak_tm, outbreak_sp] <- 5
wide_cases[c(3, 4), c(8, 9)] <- 2.5
wide_baseline <- matrix(2, nrow = 4, ncol = 9)
scanres <- parallel_shewhart_gaussian(wide_cases, wide_baseline,
                                      mean = 1,
                                      sigma = 0.6)

v_mean <- c(rep(0.0, 8), 2)
m_sigma <- matrix(1.5, nrow = 4, ncol = 9)
m_sigma[c(3, 4), c(4, 5)] <- 0.25
scanres <- parallel_shewhart_gaussian(wide_cases, wide_baseline,
                                      mean = v_mean,
                                      sigma = m_sigma)
}
