% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel2.R
\name{parallel_shewhart_gaussian2}
\alias{parallel_shewhart_gaussian2}
\title{Calculate the Gaussian Shewhart alarm statistic over multiple spatial regions}
\usage{
parallel_shewhart_gaussian2(
  spacetime_data,
  outcome_col,
  baseline_col = NULL,
  mean = 0,
  sigma = 1
)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{outcome_col}{The column in \code{spacetime_data} containing the outcome of interest, such as
observed counts}

\item{baseline_col}{Optional. The column in \code{data} containing baseline estimates of the outcome
of interest. If not present, assumed to be 0 .}

\item{mean}{The estimated mean of the residuals (\code{spacetime_data$outcome_col - spacetime_data$baseline_col}). Can either be a scalar value or a column in \code{spacetime_data}}

\item{sigma}{The estimated standard deviation of the residuals. Can either be a scalar value or a
column in \code{spacetime_data}}
}
\value{


\code{spacetime_data} with the added column \code{.action_level} containing the value of the
Shewhart statistic at that point and time.
}
\description{
\code{parallel_shewhart_gaussian} calculates the value of the Shewhart alarm statistic assuming a
Gaussian distribution independently over each spatial region, over all time points. Each spatial
region and time point is considered independently.
}
\details{
The Shewhart alarm statistic is equal to the number of standard deviations that a given
observation is above or below the mean, also known as a z-score. It can be directly calculated
as \code{(cases - baseline - mean) / sigma}.
}
\examples{
data(NM_data)

parallel_shewhart_gaussian2(NM_data, count)
parallel_shewhart_gaussian2(NM_data, count, baseline_est)
parallel_shewhart_gaussian2(NM_data, count,
                            population * mean(count) / mean(population),
                            sigma = 2)
}
\seealso{
parallel_shewhart_gaussian

Other parallel2_alarms: 
\code{\link{parallel_cusum_gaussian2}()},
\code{\link{parallel_cusum_poisson2}()}
}
\concept{parallel2_alarms}
