% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_helpers.R
\name{pivot_for_scan}
\alias{pivot_for_scan}
\title{Turn a long data frame into a wide matrix}
\usage{
pivot_for_scan(spacetime_data, value_col)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{value_col}{An expression defining the values in in \code{spacetime_data} to be pivoted into the
wide matrix. This argument uses
\href{https://dplyr.tidyverse.org/reference/dplyr_data_masking.html}{data masking} and should
be specified as bare code (not a string). Can also be a scalar, in which case all entries in
the wide matrix will be equal to \code{value_col}.}
}
\value{
A matrix, with rows corresponding to values of \code{id_time} and columns corresponding to
\code{id_space}. Exception: if \code{value_col = NULL}, then returns NULL.
}
\description{
A friendly wrapper for \code{\link[tidyr]{pivot_wider}} that takes a long data.frame
and transforms it into a wide matrix. This is the format expected as input for the #'
\code{\link[scanstatistics]{scanstatistics}} package.
}
\details{
Functions for calculating scan statistics in the
\code{\link[scanstatistics]{scanstatistics}} package expect case counts and baseline estimates
to be formatted as matrices, where the rows correspond to time (organized from earliest to
latest) and the columns correspond to space.

\code{pivot_for_scan} takes \code{spacetime_data}, and pivots the data into that format. The rows are
ordered by the \code{id_time} column from smaller to largest, and the columns are ordered by the
\code{id_space} column in \code{spacetime_data}, also from smallest to largest. \code{pivot_for_scan} expects
the data to be complete, so that every combination of \code{id_space} and \code{id_time} is associated with
exactly one value. If some combination of \code{id_space} and \code{id_time} has no corresponding value in
\code{spacetime_data}, then that entry in the returned matrix will be \verb{<NA>}.
}
\examples{
df <- data.frame(vals = c(1, 2, 3, 4, 5, 6),
                 id_space = rep(1:3, 2),
                 id_time = rep(1:2, each = 3))

pivot_for_scan(df, vals)


# Non-exist combinations create NA's
incomplete_df <- df[1:5, ]
pivot_for_scan(incomplete_df, vals)

# Returns NULL:
pivot_for_scan(df, NULL)
}
