% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_tidy.R
\name{predict.arima_tidy}
\alias{predict.arima_tidy}
\title{Forecast from arima_tidy fits}
\usage{
\method{predict}{arima_tidy}(object, only_newdata = NULL, n.ahead, se.fit = TRUE, ...)
}
\arguments{
\item{object}{The result of an \code{arima_tidy} fit}

\item{only_newdata}{New data to be used for prediction. Will be transformed into \code{newxreg} using
the formula stored in \code{object}.}

\item{n.ahead}{The number of steps ahead for which prediction is required. Optional if
\code{only_newdata} is specified, otherwise required.}

\item{se.fit}{Logical: should standard errors of prediction be returned?}

\item{...}{arguments passed to or from other methods.}
}
\value{
A time series of predictions, or if \code{se.fit = TRUE}, a list with timeseries \code{pred} and
\code{se}.
}
\description{
Forecast from models fit by \code{\link{arima_tidy}}. Calls \code{\link[stats]{predict.Arima}}
under the hood.
}
\details{
Even though \code{arima_tidy} models are fit like \code{lm} or \code{glm} models, it's important to remember
that they behave differently, because the order of the data points is crucial for all ARIMA
models. The rows in \code{only_newdata} are assumed to be ordered from earliest to latest, and are
assumed to follow immeadiately after the data used to fit the model. To extract fitted values and
generate predictions in a single step, use \code{\link{augment.arima_tidy}}.
}
\examples{
data(presidents_df)
# Fit on the first 110 rows, predict the last 10
fit <- arima_tidy(approval_rating ~ 1,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
predict(fit, n.ahead = 10)

fit <- arima_tidy(approval_rating ~ election_year,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
predict(fit, only_newdata = presidents_df[111:120, ], n.ahead = 10)
predict(fit, only_newdata = presidents_df[111:120, ])
}
\seealso{
\code{\link[stats]{arima}}, \code{\link[stats]{predict.Arima}},
\code{\link{augment.arima_tidy}}
}
