% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presidents_df}
\alias{presidents_df}
\title{Quarterly Approval Ratings of US Presidents}
\format{
A data frame with 120 rows and 3 columns:
\describe{
  \item{approval_rating}{The quarterly approval rating of the US President, as a percent (23-87)}
  \item{date}{The year, as a decimal (1945.25 - 1975.00)}
  \item{election_year}{1 if the current year was presidential election year, 0 otherwise (0/1)}
}
}
\source{
\code{\link[datasets]{presidents}}
}
\usage{
presidents_df
}
\description{
A data frame containing the approximately quarterly approval rating for the
  President of the United States, from 1945 - 1974. Converted to a data frame from the
  \code{\link[datasets]{presidents}} time series.
}
\keyword{datasets}
