% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_top_clusters.R
\name{remove_overlapping_clusters}
\alias{remove_overlapping_clusters}
\title{Remove overlapping clusters from scan statistics}
\usage{
remove_overlapping_clusters(
  scan_result,
  zones,
  order_by,
  take_highest = TRUE,
  max_overlap_frac = 0
)
}
\arguments{
\item{scan_result}{An object of class \code{scanstatistic}, such as generated by
\code{\link[scanstatistics]{scan_bayes_negbin}} or \code{\link{scan_eb_poisson2}}.}

\item{zones}{A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.}

\item{order_by}{The column to order the clusters by. Any column in \code{scan_result$observed} or
\code{scan_result$posteriors$window_posteriors} is valid. Specified as bare text.}

\item{take_highest}{Boolean. Should the clusters by ordered from highest to lowest (\code{take_highest = TRUE}) or from lowest to highest (\code{take_highest = FALSE}).}

\item{max_overlap_frac}{Optional. What is the maximum overlap fraction that should be allowed
between clusters? Defaults to 0.}
}
\value{
An object of class \code{scanstatistic}, with the \code{observed} or \code{posteriors$window_posteriors}
data frame modified to only contain non-overlapping clusters. Otherwise identical to the
\code{scan_result} input.
}
\description{
In addition to reporting the details of the most likely cluster (MLC),
\code{scanstatistics} objects also contain detailed information about the likelihood of a cluster in
all possible spatial zones and temporal durations. This information can be used to identify any
independent secondary clusters that exist in the study area. When searching for these secondary
clusters, we are generally interested in clusters that do not significantly overlap the primary
cluster.
}
\details{
\code{remove_overlapping_clusters} goes through the information about the likelihood of a
cluster in all possible spatial zones and temporal durations, and only keeps the information
from non-overlapping spatial zones. Potential clusters are gone through from most-likely to
least likely (as determined by \code{order_by} and \code{take_highest}). Information about a spatial zone
and duration is retained if the fraction of spatial locations in the zone that are present in
any already identified cluster is less than or equal to \code{max_overlap_frac}.
}
\examples{
library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scanres <- scan_eb_poisson2(NM_data, count, zones,
                            baseline_est, n_mcsim = 100)

# Order by cluster score
remove_overlapping_clusters(scanres, zones, order_by = score)

# Order by mc-pvalue
remove_overlapping_clusters(scanres, zones, order_by = mc_pvalue,
                            take_highest = FALSE)

# Order by mc-pvalue, allow some overlap
remove_overlapping_clusters(scanres, zones, order_by = mc_pvalue,
                            take_highest = FALSE,
                            max_overlap_frac = 0.2)
}
\seealso{
report_top_clusters
}
