% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model_coeff.R
\name{report_model_}
\alias{report_model_}
\alias{report_model_coeff}
\alias{report_model_summary}
\title{Return information about a model or its coefficients}
\usage{
report_model_coeff(fit, include_random = FALSE)

report_model_summary(fit)
}
\arguments{
\item{fit}{The model fit object. The model fit can be generated by INLA or by any model function
recognized by the \code{\link[broom]{broom}} package.}

\item{include_random}{In \code{report_model_coeff}, for INLA models only, should random effects be
included in the returned data.frame? If FALSE, only fixed effects and model hyper-parameters
will be returned.}
}
\value{
For \code{report_model_coeff}, a \code{\link[tibble]{tibble}} containing estimates and
standard errors of the coefficients in \code{fit}. For \code{report_model_summary}, a 1-row summary
tibble containing information of goodness of fit, p-values, or model convergence.
}
\description{
\code{report_model_coeff} and \code{report_model_summary} are simple wrappers around the
functions \code{\link[broom]{tidy}} and \code{\link[broom]{glance}} from the broom package that
can also report information from \code{\link[INLA]{inla}} models.
}
\examples{
x <- rnorm(20, mean = 0, sd = 3)
y <- 1.3 + 2.5 * x + rnorm(20, sd = 1)
df <- data.frame(x = x, y = y)


fit_lm <- lm(y ~ x, data = df)
report_model_coeff(fit_lm)
report_model_summary(fit_lm)

\dontrun{

fit_inla <- INLA::inla(y ~ x, data = df)
report_model_coeff(fit_inla)
report_model_summary(fit_inla)
}

}
\seealso{
\code{\link[broom]{tidy}}, \code{\link[broom]{glance}}
}
