% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_top_clusters.R
\name{report_top_clusters}
\alias{report_top_clusters}
\title{Extract clusters identified by scanning alarm functions}
\usage{
report_top_clusters(
  scan_result,
  order_by,
  take_highest = TRUE,
  cutoff_val = NULL,
  max_reported = 10,
  min_reported = 1,
  single_duration = TRUE
)
}
\arguments{
\item{scan_result}{An object of class \code{scanstatistic}, such as generated by
\code{\link[scanstatistics]{scan_bayes_negbin}} or \code{\link{scan_eb_poisson2}}.}

\item{order_by}{The column to order the clusters by. Any column in \code{scan_result$observed} or
\code{scan_result$posteriors$window_posteriors} is valid. Specified as bare text.}

\item{take_highest}{Boolean. Should the clusters by ordered from highest to lowest (\code{take_highest = TRUE}) or from lowest to highest (\code{take_highest = FALSE}).}

\item{cutoff_val}{The cutoff value used for reporting clusters. If \code{take_highest = TRUE}, only
clusters with values of \code{order_by} greater than \code{cutoff_val} will be reported; if
\code{take_highest = FALSE}, only report clusters with values under \code{cutoff_val}.
Defaults to \code{NULL}, where no cutoff value is applied.}

\item{max_reported}{The maximum number of clusters to be reported. Defaults to 10.}

\item{min_reported}{The minimum number of clusters to be reported. This number of clusters will
always be reported, even if they do not meet the threshold set by \code{cutoff_val}.}

\item{single_duration}{Boolean. If TRUE, remove clusters with the same spatial zone but different
duration as another, more intense, cluster.}
}
\value{
A data frame containing information about the most likely clusters in the study area,
with between \code{min_reported} and \code{max_reported} rows, and including columns \code{zone}, \code{duration},
and the column specified by \code{order_by}.
}
\description{
Extract all space-time clusters, based on values calculated by a scan statistic.
This can either be an absolute measure of intensity, such as the \code{score} calculate by
\code{\link[scanstatistics]{scan_eb_poisson}} or a relative measure, such as the Monte-Carlo
p-value.
}
\details{
\code{scanstatistics} objects contain detailed information about the likelihood of a cluster
in all possible spatial zones and temporal durations. \code{report_top_clusters} uses this
information to report the zones most likely to contain a cluster. The results can be customized
using the \code{cutoff_val}, \code{max_reported}, \code{min_reported}, and \code{single_duration} options.
}
\examples{
library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scanres <- scan_eb_poisson2(NM_data, count, zones,
                            baseline_est, n_mcsim = 100)

report_top_clusters(scanres, score)

report_top_clusters(scanres, mc_pvalue,
                    take_highest = FALSE,
                    cutoff_val = 0.05)

# Always report at least 3 clusters:
report_top_clusters(scanres, mc_pvalue,
                    take_highest = FALSE,
                    cutoff_val = 0.05,
                    min_reported = 3)

# First remove overlapping clusters, then report:
remove_overlapping_clusters(scanres, zones,
                            order_by = mc_pvalue,
                            take_highest = FALSE,
                            max_overlap_frac = 0.2) \%>\%
  report_top_clusters(mc_pvalue, take_highest = FALSE,
                      cutoff_val = 0.05)
}
\seealso{
remove_overlapping_clusters
}
