% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_statistics2.R
\name{scan_bayes_negbin2}
\alias{scan_bayes_negbin2}
\title{Calculate the negative binomial bayesian scan statistic..}
\usage{
scan_bayes_negbin2(
  spacetime_data,
  outcome_col,
  zones,
  baseline_col = NULL,
  pop_col = NULL,
  outbreak_prob = 0.05,
  alpha_null = 1,
  beta_null = 1,
  alpha_alt = alpha_null,
  beta_alt = beta_null,
  inc_values = seq(1, 3, by = 0.1),
  inc_probs = 1
)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{outcome_col}{The column in \code{spacetime_data} containing the outcome of interest, such as
observed counts}

\item{zones}{A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.}

\item{baseline_col}{The column in data containing baseline estimates of the outcome of interest.
Optional if \code{pop_col} is provided, otherwise required.}

\item{pop_col}{Optional. The column in \code{spacetime_data} containing the population for each
location/time. Used to estimate the baseline if \code{baseline_col} is NULL, otherwise ignored.}

\item{outbreak_prob}{A scalar; the probability of an outbreak (at any time,
any place). Defaults to 0.05.}

\item{alpha_null}{A scalar; the shape parameter for the gamma distribution
under the null hypothesis of no anomaly. Defaults to 1.}

\item{beta_null}{A scalar; the scale parameter for the gamma distribution
under the null hypothesis of no anomaly. Defaults to 1.}

\item{alpha_alt}{A scalar; the shape parameter for the gamma distribution
under the alternative hypothesis of an anomaly. Defaults to the same value
as \code{alpha_null}.}

\item{beta_alt}{A scalar; the scale parameter for the gamma distribution
under the alternative hypothesis of an anomaly. Defaults to the same value
as \code{beta_null}.}

\item{inc_values}{A vector of possible values for the increase in the mean
(and variance) of an anomalous count. Defaults to evenly spaced values 
between 1 and 3, with a difference of 0.1 between consecutive values.}

\item{inc_probs}{A vector of the prior probabilities of each value in 
\code{inc_values}. Defaults to 1, implying a discrete uniform 
distribution.}
}
\value{
An object of class \code{scanstatistics}, containing information about the type of scan
statistic, the most likely cluster (MLC), and calculated scan statistics for every zone and
duration investigated.
}
\description{
Calculate the "Bayesian Spatial Scan Statistic" by Neill et al. (2006), adapted to a
spatio-temporal setting. \code{scan_bayes_negbin2} is a  helpful wrapper around
\code{\link[scanstatistics]{scan_bayes_negbin}} allowing data to be passed as a single long
data frame, instead of as a series of wide matrices.
}
\examples{
library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scan_bayes_negbin2(NM_data, outcome_col = count, zones,
                   baseline_col = baseline_est)

scan_bayes_negbin2(NM_data, outcome_col = count, zones,
                   pop_col = population)

scan_bayes_negbin2(NM_data, outcome_col = count, zones,
                   baseline_col = baseline_est,
                   outbreak_prob = 0.1)

}
\seealso{
\code{\link[scanstatistics]{scan_bayes_negbin}}

Other v2 scan statistics: 
\code{\link{scan_cusum_poisson2}()},
\code{\link{scan_eb_negbin2}()},
\code{\link{scan_eb_negbin_fast2}()},
\code{\link{scan_eb_poisson2}()},
\code{\link{scan_eb_poisson_fast2}()},
\code{\link{scan_eb_zip2}()},
\code{\link{scan_pb_poisson2}()},
\code{\link{scan_permutation2}()}
}
\concept{v2 scan statistics}
