% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_cusum_poisson.R
\name{scan_cusum_poisson}
\alias{scan_cusum_poisson}
\title{Calculate the CUSUM statistic over a set of spatial-temporal zones}
\usage{
scan_cusum_poisson(
  wide_cases,
  key_matrix,
  wide_baseline,
  scaling = 1.5,
  n_mcsim = 100
)
}
\arguments{
\item{wide_cases}{A matrix of observed counts. Rows indicate time and are ordered from least
recent (row 1) to most recent (row \code{nrow(counts)}). Columns indicate locations, numbered from 1
and up. Can be generated from a data.frame with the \code{\link{pivot_for_scan}} function.}

\item{key_matrix}{A matrix specifying how which locations belong to each zone. Each row is a
different zone and each column is a location. An element \verb{(i,j)} is 1 if location \code{j} is part
of zone \code{i} and is 0 otherwise. Can be generated from a list specifying zones using the
\code{\link{zones_to_key_matrix}} function.}

\item{wide_baseline}{A matrix of the same dimensions as \code{wide_cases}, holding the expected value
of cases at each space-time location.This value is typically estimated from past data using a
GLM or other model.}

\item{scaling}{A scalar number > 1. Represents the proportional increase in the number of cases
beyond \code{wide_baseline} in the event of an outbreak. If \code{wide_baseline} is the number of
expected cases under normal conditions, then \code{scaling * wide_baseline} is the number of
expected cases in the event of an outbreak. Used to calculate the downward drift.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan
statistics to generate in order to calculate a \eqn{P}-value.}
}
\value{
A list which, in addition to the information about the type of scan
   statistic, has the following components:
   \describe{
     \item{MLC}{A list containing the number of the zone of the most likely
           cluster (MLC), the locations in that zone, the duration of the 
           MLC, the calculated score, and the relative risk. In order, the 
           elements of this list are named  \code{zone_number, locations, 
           duration, score, relative_risk}.}
     \item{observed}{A data frame containing, for each combination of zone 
           and duration investigated, the zone number, duration, score, 
           relative risk. The table is sorted by score with the top-scoring 
           location on top. If \code{max_only = TRUE}, only contains a single 
           row corresponding to the MLC.}
     \item{replicates}{A data frame of the Monte Carlo replicates of the scan 
           statistic (if any), and the corresponding zones and durations.}
     \item{MC_pvalue}{The Monte Carlo \eqn{P}-value.}
     \item{Gumbel_pvalue}{A \eqn{P}-value obtained by fitting a Gumbel 
           distribution to the replicate scan statistics.}
     \item{n_zones}{The number of zones scanned.}
     \item{n_locations}{The number of locations.}
     \item{max_duration}{The maximum duration considered.}
     \item{n_mcsim}{The number of Monte Carlo replicates made.}
   }
}
\description{
Calculate the CUSUM alarm statistic over a large set of spatio-temporal zones, as
described by Sonesson (2007).
}
\details{
The CUSUM statistic is widely used surveillance technique for purely temporal processes.
It is calculated by taking the cumulative sum of the standard residuals with a downward drift
and a floor of 0. For a Poisson process \eqn{x(t)} with expected value \eqn{\lambda} and an
expected value during an ourbreak of \eqn{\delta\lambda} the optimal downward drift is \eqn{k =
  \lambda(\delta - 1) / (log(\delta)}.

When applied as a scan statistic, the total number of cases and the expected number of cases are
aggregated for a large number of spatial-temporal zones, where each zone consists of one or more
neighboring spatial locations.
For each zone, a single CUSUM statistic is calculated, incorporating information from the
earliest time point to the most recent. In the returned value, all \code{duration} fields refer to
the number of time points since the CUSUM statistic was last equal to 0. This is different from
the behavoir of the other scan-type functions, where the duration indicates the size of the
temporal window.
}
\examples{
# Apply scan_cusum_poisson to a 3x3 spatial grid with 4 time points
# And inject an outbreak into the upper-left corner in the 2 most recent time points
library("magrittr")
x_coord <- rep(c(1:3), 3)
y_coord <- rep(c(1:3), each =3)
geo <- matrix(c(x_coord, y_coord), ncol = 2, byrow = FALSE)
zones <- geo \%>\%
  scanstatistics::coords_to_knn(k = 4) \%>\%
  scanstatistics::knn_zones()
key_matrix <- zones_to_key_matrix(zones)
outbreak_sp <- c(1, 2, 4, 5)
outbreak_tm <- c(3, 4)
wide_cases <- matrix(2, nrow = 4, ncol = 9)
wide_cases[outbreak_tm, outbreak_sp] <- 5
wide_cases[c(3, 4), c(8, 9)] <- 2.5
wide_baseline <- matrix(2, nrow = 4, ncol = 9)
scanres <- scan_cusum_poisson(wide_cases, key_matrix, wide_baseline,
                              scaling = 1.5,
                              n_mcsim = 10)
}
\references{
Sonesson, C. (2007). \emph{A CUSUM framework for detection of space-time disease clusters using
scan statistics}, Stat Med 26(26): 4770-4789.
}
\seealso{
\code{\link{parallel_cusum_poisson}}, \code{\link{pivot_for_scan}},
\code{\link{zones_to_key_matrix}}
}
