% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_statistics2.R
\name{scan_cusum_poisson2}
\alias{scan_cusum_poisson2}
\title{Calculate the CUSUM statistic over a set of spatial-temporal zones}
\usage{
scan_cusum_poisson2(
  spacetime_data,
  outcome_col,
  zones,
  baseline_col,
  scaling = 1.5,
  n_mcsim = 10
)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{outcome_col}{The column in \code{spacetime_data} containing the outcome of interest, such as
observed counts}

\item{zones}{A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.}

\item{baseline_col}{The column in data containing baseline estimates of the outcome of interest.
Optional if \code{pop_col} is provided, otherwise required.}

\item{scaling}{A scalar number > 1. Represents the proportional increase in the number of cases
beyond \code{wide_baseline} in the event of an outbreak. If \code{wide_baseline} is the number of
expected cases under normal conditions, then \code{scaling * wide_baseline} is the number of
expected cases in the event of an outbreak. Used to calculate the downward drift.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan
statistics to generate in order to calculate a \eqn{P}-value.}
}
\value{
An object of class \code{scanstatistics}, containing information about the type of scan
statistic, the most likely cluster (MLC), and calculated scan statistics for every zone and
duration investigated.
}
\description{
Calculate the CUSUM alarm statistic over a large set of spatio-temporal zones, as
described by Sonesson (2007). \code{scan_cusum_poisson2} is a helpful wrapper around
\code{\link{scan_cusum_poisson}} allowing data to be passed as a single long data frame,
instead of as a series of wide matrices.
}
\examples{
library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scan_cusum_poisson2(NM_data, outcome_col = count, zones,
                    baseline_col = baseline_est)

scan_cusum_poisson2(NM_data, outcome_col = count, zones,
                    baseline_col = baseline_est,
                    scaling = 1.25,
                    n_mcsim = 99)

}
\seealso{
\code{\link{scan_cusum_poisson}} \code{\link{parallel_cusum_poisson}}

Other v2 scan statistics: 
\code{\link{scan_bayes_negbin2}()},
\code{\link{scan_eb_negbin2}()},
\code{\link{scan_eb_negbin_fast2}()},
\code{\link{scan_eb_poisson2}()},
\code{\link{scan_eb_poisson_fast2}()},
\code{\link{scan_eb_zip2}()},
\code{\link{scan_pb_poisson2}()},
\code{\link{scan_permutation2}()}
}
\concept{v2 scan statistics}
