% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_statistics2.R
\name{scan_eb_poisson2}
\alias{scan_eb_poisson2}
\title{Calculate the expectation-based Poisson scan statistic.}
\usage{
scan_eb_poisson2(
  spacetime_data,
  outcome_col,
  zones,
  baseline_col = NULL,
  pop_col = NULL,
  n_mcsim = 10,
  gumbel = FALSE,
  max_only = FALSE
)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{outcome_col}{The column in \code{spacetime_data} containing the outcome of interest, such as
observed counts}

\item{zones}{A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.}

\item{baseline_col}{The column in data containing baseline estimates of the outcome of interest.
Optional if \code{pop_col} is provided, otherwise required.}

\item{pop_col}{Optional. The column in \code{spacetime_data} containing the population for each
location/time. Used to estimate the baseline if \code{baseline_col} is NULL, otherwise ignored.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan
statistics to generate in order to calculate a \eqn{P}-value.}

\item{gumbel}{Logical: should a Gumbel P-value be calculated? Default is
\code{FALSE}.}

\item{max_only}{Boolean. If \code{FALSE} (default) the log-likelihood ratio
statistic for each zone and duration is returned. If \code{TRUE}, only the
largest such statistic (i.e. the scan statistic) is returned, along with
the corresponding zone and duration.}
}
\value{
An object of class \code{scanstatistics}, containing information about the type of scan
statistic, the most likely cluster (MLC), and calculated scan statistics for every zone and
duration investigated.
}
\description{
Calculate the expectation-based Poisson scan statistic using data frames.
\code{scan_eb_poisson2} is a helpful wrapper around \code{\link[scanstatistics]{scan_eb_poisson}}
allowing data to be passed as a single long data frame, instead of as a series of wide
matrices.
}
\examples{
library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scan_eb_poisson2(NM_data, outcome_col = count, zones,
                 baseline_col = baseline_est, n_mcsim = 100)
scan_eb_poisson2(NM_data, outcome_col = count, zones,
                 pop_col = population, n_mcsim = 100, gumbel = TRUE)
}
\seealso{
\code{\link[scanstatistics]{scan_eb_poisson}}

Other v2 scan statistics: 
\code{\link{scan_bayes_negbin2}()},
\code{\link{scan_cusum_poisson2}()},
\code{\link{scan_eb_negbin2}()},
\code{\link{scan_eb_negbin_fast2}()},
\code{\link{scan_eb_poisson_fast2}()},
\code{\link{scan_eb_zip2}()},
\code{\link{scan_pb_poisson2}()},
\code{\link{scan_permutation2}()}
}
\concept{v2 scan statistics}
