% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_eb_poisson_fast.R
\name{scan_eb_poisson_fast}
\alias{scan_eb_poisson_fast}
\title{Calculate the expectation-based Poisson scan statistic via linear algebra}
\usage{
scan_eb_poisson_fast(wide_cases, key_matrix, wide_baseline, n_mcsim)
}
\arguments{
\item{wide_cases}{A matrix of observed counts. Rows indicate time and are ordered from least
recent (row 1) to most recent (row \code{nrow(counts)}). Columns indicate locations, numbered from 1
and up. Can be generated from a data.frame with the \code{\link{pivot_for_scan}} function.}

\item{key_matrix}{A matrix specifying how which locations belong to each zone. Each row is a
different zone and each column is a location. An element \verb{(i,j)} is 1 if location \code{j} is part
of zone \code{i} and is 0 otherwise. Can be generated from a list specifying zones using the
\code{\link{build_key_matrix}} function.}

\item{wide_baseline}{A matrix of the same dimensions as \code{wide_cases}, holding the expected value
of cases at each space-time location.This value is typically estimated from past data using a
GLM or other model.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan
statistics to generate in order to calculate a \eqn{P}-value.}
}
\value{
A list which, in addition to the information about the type of scan
   statistic, has the following components:
   \describe{
     \item{MLC}{A list containing the number of the zone of the most likely
           cluster (MLC), the locations in that zone, the duration of the 
           MLC, the calculated score, and the relative risk. In order, the 
           elements of this list are named  \code{zone_number, locations, 
           duration, score, relative_risk}.}
     \item{observed}{A data frame containing, for each combination of zone 
           and duration investigated, the zone number, duration, score, 
           relative risk. The table is sorted by score with the top-scoring 
           location on top. If \code{max_only = TRUE}, only contains a single 
           row corresponding to the MLC.}
     \item{replicates}{A data frame of the Monte Carlo replicates of the scan 
           statistic (if any), and the corresponding zones and durations.}
     \item{MC_pvalue}{The Monte Carlo \eqn{P}-value.}
     \item{Gumbel_pvalue}{A \eqn{P}-value obtained by fitting a Gumbel 
           distribution to the replicate scan statistics.}
     \item{n_zones}{The number of zones scanned.}
     \item{n_locations}{The number of locations.}
     \item{max_duration}{The maximum duration considered.}
     \item{n_mcsim}{The number of Monte Carlo replicates made.}
   }
}
\description{
Calculate the expectation-based scan statistic using the a Poisson distribution, as
devised by Neill (2009). This function should give identical results to those calculated by
\code{\link[scanstatistics]{scan_eb_poisson}}, but is significantly faster when scanning large
areas.
}
\examples{
# Apply scan_eb_poisson_fast to a 3x3 spatial grid with 4 time points
# And inject an outbreak into the upper-left corner in the 2 most recent time points
library("magrittr")
x_coord <- rep(c(1:3), 3)
y_coord <- rep(c(1:3), each =3)
geo <- matrix(c(x_coord, y_coord), ncol = 2, byrow = FALSE)
zones <- geo \%>\%
  scanstatistics::coords_to_knn(k = 4) \%>\%
  scanstatistics::knn_zones()
key_matrix <- build_key_matrix(zones)
outbreak_sp <- c(1, 2, 4, 5)
outbreak_tm <- c(3, 4)
wide_cases <- matrix(2, nrow = 4, ncol = 9)
wide_cases[outbreak_tm, outbreak_sp] <- 5
wide_cases[c(3, 4), c(8, 9)] <- 2.5
wide_baseline <- matrix(2, nrow = 4, ncol = 9)
scanres <- scan_eb_poisson_fast(wide_cases, key_matrix, wide_baseline,
                                n_mcsim = 10)
}
\references{
Neill, D. B., Moore, A. W., Sabhnani, M. and Daniel, K. (2005). 
   \emph{Detection of emerging space-time clusters}. Proceeding of the 
   eleventh ACM SIGKDD international conference on Knowledge discovery in 
   data mining - KDD ’05, 218.
}
\seealso{
\code{\link[scanstatistics]{scan_eb_poisson}}, \code{\link{scan_eb_negbin_fast}},
\code{\link{pivot_for_scan}}, \code{\link{build_key_matrix}}
}
