% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_helpers.R
\name{space_coord_to_zones}
\alias{space_coord_to_zones}
\title{Create zones from spatial coordinates}
\usage{
space_coord_to_zones(space_coord, max_k)
}
\arguments{
\item{space_coord}{Either a SpatialPolygonsDataFrame or a simple-features (sf) collection
describing the spatial locations to combine into zones.}

\item{max_k}{The maximum number of locations to be included in any one zone.}
}
\value{
A list of integer vectors, specifying which locations belong to which zones. The integers
identify locations based on which row in \code{space_coord} they are associated with.
}
\description{
Scan-type alarm functions work by computing an alarm function over a very large
number of spatial \emph{zones}, where each zone consists of one of more neighboring spatial
locations. Zones can be many sizes, from a single location up to an aggregation of \code{max_k}
locations, and each location can belong to multiple zones. \code{space_coord_to_zones} aggregates
individual spatial locations into a set of zones using a k-nearest neighbors approach. For each
location, there will be a zone containing the location itself, a zone containing the location
and its nearest neighbor, a zone containing the location and its two nearest-neighbors, up to
the maximum size.
}
\examples{
library("sf")
space_coord_to_zones(co_county_fips_2019, 5)
# 297 zones created from 64 locations
}
