% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacetime_data.R
\name{spacetime_data}
\alias{spacetime_data}
\title{Data frames including information on space and time}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}
}
\description{
Data frames including information on space and time
}
