% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_alarm_functions.R
\name{standardized_alarm_functions}
\alias{standardized_alarm_functions}
\title{A standardized interface for calling alarm functions}
\usage{
standardized_alarm_functions(
  alarm_function_name,
  wide_cases,
  zone_info,
  wide_baseline,
  ...,
  n_mcsim
)
}
\arguments{
\item{alarm_function_name}{The name of the alarm function to call. See the list of supported
alarm functions below.}

\item{wide_cases}{A matrix of observed counts. Rows indicate time and are ordered from least
recent (row 1) to most recent (row \code{nrow(counts)}). Columns indicate locations, numbered from 1
and up. Can be generated from a data.frame with the \code{\link{pivot_for_scan}} function.}

\item{zone_info}{Either:
\itemize{
\item A list of integer vectors. Each vector corresponds to a single
zone; its elements are the numbers of the locations in that zone.
\item A matrix specifying how which locations belong to each zone. Each row is a different
zone and each column is a location. An element \verb{(i,j)} is 1 if location \code{j} is part of zone
\code{i} and is 0 otherwise. Can be generated from a list specifying zones using the
\code{\link{build_key_matrix}} function.
}}

\item{wide_baseline}{A matrix of the same dimensions as counts, holding the expected value for
each observed space-time location. This value is typically estimated from past data using a GLM
or other model. This parameter is passed into the \code{population} argument for
\code{\link[scanstatistics]{scan_pb_poisson}} and
\code{\link[scanstatistics]{scan_permutation}}, and into the \code{baselines} or \code{wide_baselines}
argument for all other alarm functions.}

\item{...}{Additional parameters to pass to the specified alarm function.}

\item{n_mcsim}{A non-negative integer; the number of replicate scan
statistics to generate in order to calculate a \eqn{P}-value.}
}
\value{
For the alarm functions that begin "parallel_", a matrix with the same dimensions as
\code{wide_cases}; for all other alarm functions an object of class \code{scanstatistics}. All results
have an attribute "alarm_type" describing whether they are a scan-type alarm function or a
parallel-type alarm function. All returned \code{scanstatistics} objects have been
standardized as follows:
\itemize{
\item{The returned object now contains the component \code{zone_info} containing information about
the spatio-temporal zones}
\item{The returned object always contains the component \code{observed}, giving the likelihood of an
outbreak in each spatio-temporal zone},
\item{The components \code{observed}, \code{MLC}, and \code{replicates} (if applicable) always contain the
column \code{action_level}. The higher the value of this column, the greater the likelihood of an
outbreak. This column is identical to \code{score} for all scan-type alarm functions except for
\code{\link[scanstatistics]{scan_bayes_negbin}}, where it is equal to \code{log_posterior}}
}
}
\description{
Call any of 12 alarm functions with a consistent interface.
}
\section{Supported Alarm Functions}{

The following alarm functions are supported:
\itemize{
\item{\code{\link[scanstatistics]{scan_eb_poisson}}}
\item{\code{\link[scanstatistics]{scan_pb_poisson}}}
\item{\code{\link[scanstatistics]{scan_eb_negbin}}}
\item{\code{\link[scanstatistics]{scan_eb_zip}}}
\item{\code{\link[scanstatistics]{scan_permutation}}}
\item{\code{\link[scanstatistics]{scan_bayes_negbin}}}
\item{\code{\link{scan_eb_poisson_fast}}}
\item{\code{\link{scan_eb_negbin_fast}}}
\item{\code{\link{scan_cusum_poisson}}}
\item{\code{\link{parallel_cusum_poisson}}}
\item{\code{\link{parallel_cusum_gaussian}}}
\item{\code{\link{parallel_shewhart_gaussian}}}
}
}

\examples{
# Apply scan_cusum_poisson to a 3x3 spatial grid with 4 time points
# Inject an outbreak into the upper-left corner in
# the 2 most recent time points
library("magrittr")
x_coord <- rep(c(1:3), 3)
y_coord <- rep(c(1:3), each =3)
geo <- matrix(c(x_coord, y_coord), ncol = 2, byrow = FALSE)
zones <- geo \%>\%
  scanstatistics::coords_to_knn(k = 4) \%>\%
  scanstatistics::knn_zones()
outbreak_sp <- c(1, 2, 4, 5)
outbreak_tm <- c(3, 4)
wide_cases <- matrix(2, nrow = 4, ncol = 9)
wide_cases[outbreak_tm, outbreak_sp] <- 5
wide_cases[c(3, 4), c(8, 9)] <- 3
wide_baseline <- matrix(2, nrow = 4, ncol = 9)

standardized_alarm_functions("scan_eb_poisson",
                             wide_cases, zones, wide_baseline,
                             n_mcsim = 10)
standardized_alarm_functions("scan_permutation",
                             wide_cases, zones, wide_baseline,
                             n_mcsim = 10)
standardized_alarm_functions("scan_eb_negbin_fast",
                             wide_cases, zones, wide_baseline,
                             thetas = 1.5, n_mcsim = 10)
}
