% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_helpers.R
\name{switch_zone_rep}
\alias{switch_zone_rep}
\alias{zones_to_key_matrix}
\alias{key_matrix_to_zones}
\title{Change how spatial zones are represented}
\usage{
zones_to_key_matrix(zones)

key_matrix_to_zones(key_matrix)
}
\arguments{
\item{zones}{A list where each entry is a zone. Each entry is a vector containing the indices of
the spatial locations in that zone.}

\item{key_matrix}{A matrix, where each row is a zone and each column is a location. The entry
\verb{[i, j]} is 1 if location \code{j} is a part of zone \code{i} and 0 otherwise.}
}
\value{
For \code{zones_to_key_matrix}, a matrix. To increase speed, the matrix is represented as a
sparse matrix from the \code{\link[Matrix]{Matrix}} package. For \code{create_zones}, and
\code{key_matrix_to_zones} a list of zones.
}
\description{
Scan-type alarm functions work by computing an alarm function over a very large
number of spatial \emph{zones}, where each zone consists of one of more neighboring spatial
locations. Information about which spatial locations belong to which zone can be specified in
the form of a list, where each entry in the list contains the indexes of the locations in a
particular zone, or in the form of a matrix, where each row represents a zone and each column
represents a location. A set of zones can be generated using the
\code{\link{create_zones}} function, and conversion between the list- and
matrix-representation of the zones is accomplished through the functions
\code{zones_to_key_matrix} and \code{key_matrix_to_zones}.
}
\examples{
library("sf")
data("NM_county_sf")

zones <- create_zones(NM_county_sf, max_k = 5)
key_matrix <- zones_to_key_matrix(zones)

zones <- key_matrix_to_zones(key_matrix)
}
\seealso{
\code{\link{create_zones}},
\code{\link[scanstatistics]{knn_zones}}, \code{\link[scanstatistics]{dist_to_knn}}
}
