% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inla.R
\name{tidy_inla}
\alias{tidy_inla}
\title{Report information about the components of an INLA model}
\usage{
tidy_inla(fit, include_random = FALSE)
}
\arguments{
\item{fit}{The model fit object generated by INLA.}

\item{include_random}{Boolean. Should information about random effects be included in the
results?}
}
\value{
A data frame containing estimates and standard errors of the coefficients in \code{fit}, with
one row per coefficient. The result will always contain the results for fixed effects and
hyperparameters. If \code{include_random = TRUE}, the result will also include all the random
effects, such as the individual effects from an \code{iid} or \code{BYM} model. The data frame contains
the columns "term", "estimate", "std.error", "statistic", "p.value", and "type".
}
\description{
\code{\link[broom]{tidy}} summarizes information about the components of a model. \code{tidy_inla}
adapts this to INLA models.
}
\examples{
x <- rnorm(20, mean = 0, sd = 3)
y <- 1.3 + 2.5 * x + rnorm(20, sd = 1)
df <- data.frame(x = x, y = y)

\dontrun{
fit_inla <- INLA::inla(y ~ x, data = df)
tidy_inla(fit_inla)

data(NM_data)
fit_inla <- INLA::inla(count ~ population + f(county, model = "iid"),
                       data = NM_data)
tidy_inla(fit_inla, include_random = TRUE)
}
}
\seealso{
\code{\link[broom]{tidy}} \code{\link{glance_inla}}
}
