% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{unlist_scalars}
\alias{unlist_scalars}
\title{Convert list columns back to regular columns where possible}
\usage{
unlist_scalars(df)
}
\arguments{
\item{df}{A data frame or data frame extension. Can be grouped.}
}
\value{
A data frame with the same column names and grouping structure as the input.
}
\description{
Given a data frame, convert any list columns containing only atomic scalars into regular columns.
Any grouping structure of the data frame is preserved.
}
\examples{
# Create a data frame with some list columns
df <- tibble::tibble(x = c(1, 2, 3),
                     y = list("A", "B", "C"),
                     z = list(14, 201, c(99, 45)))

# Converts y into a regular column, leaves z as a list column
unlist_scalars(df)
}
