% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing_functions.R
\name{window_idtime}
\alias{window_idtime}
\title{Divide data into overlapping windows}
\usage{
window_idtime(
  spacetime_data,
  min_train,
  max_train = Inf,
  n_predict = 1,
  model_arity = c("multi", "uni")
)
}
\arguments{
\item{spacetime_data}{A data frame or a data frame extension (e.g., a tibble) that includes
columns `id_space` and `id_time` that are each a set of consecutive integers. The column
`id_time` identifies the time point associated with each row, with lower numbers being earlier
and higher numbers being later. The column `id_space` identifies the spatial location or area
associated with each row, but no meaning or order is ascribed to the values.}

\item{min_train, max_train}{The minimum and maximum number of time points (as identified by
\code{id_time}) to include in the training data set. Times with fewer than this number of time
points will not be included as windows. Later times will include more time points in the
training data set, up to the maximum number identified by \code{max_train}.}

\item{n_predict}{The number of time points (as identified by \code{id_time}) to have predictions
generated for.}

\item{model_arity}{Whether each spatial area (as identified by \code{id_space}) should be located in
its own data window (if \code{model_arity = "uni"}) or if they should all be included in one window
(if \code{model_arith = "multi"}).}
}
\value{
A tibble with one row per data window, containing the columns \code{id_time}, \code{id_space},
\code{split_id}, and \code{curr_data}. \code{id_time} and \code{id_space} identify the time point defining the end
of the time point and the spatial area or areas associated with the data window.  \code{split_id} is
the first time point in the data window in the prediction set. The data for each window is
stored in the \code{curr_data} column. There is no difference in the windowed data between data in
the training set and the prediction set, but that information is used by other \code{statsurv}
functions.
}
\description{
This function divides a dataframe into overlapping time windows, in preparation for repeatedly
applying statistical surveillance methods. Each window consists of 0 or more data points to fit a
statistical model on, and 0 or more points to generate predictions for.
}
\examples{
x = rnorm(100)
spacetime_data <- data.frame(id_space = rep(1:10, each = 10),
                             id_time = rep(1:10, 10),
                             x = x,
                             y = 2.04 * x + 1.23)

Create overlapping groups of data with at least 4 time points and up to 8 time points in the
training set, and 2 time points in the prediction set.
window_idtime(spacetime_data, min_train = 4, max_train = 8,
              n_predict = 2, model_arity = "multi")

The same, but put each spatial area in its own window
window_idtime(spacetime_data, min_train = 4, max_train = 8,
              n_predict = 2, model_arity = "uni")
}
