% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstatistics_helpers.R
\name{switch_zone_rep}
\alias{switch_zone_rep}
\alias{build_key_matrix}
\alias{build_zones}
\title{Change how spatial zones are represented}
\usage{
build_key_matrix(zones)

build_zones(key_matrix)
}
\arguments{
\item{zones}{A list where each entry is a zone. Each entry is a vector containing the indices of
the spatial locations in that zone.}

\item{key_matrix}{A matrix, where each row is a zone and each column is a location. The entry
\verb{[i, j]} is 1 if location \code{j} is a part of zone \code{i} and 0 otherwise.}
}
\value{
For \code{build_key_matrix}, a matrix. To increase speed, the matrix is represented as a
sparse matrix from the \code{\link[Matrix]{Matrix}} package. For \code{build_zones}, a list of
zones.
}
\description{
Scan-type alarm functions work by computing an alarm function over a very large
number of spatial \emph{zones}, where each zone consists of one of more neighboring spatial
locations. Information about which spatial locations belong to which zone can be specified in
the form of a list, where each entry in the list contains the indexes of the locations in a
particular zone, or in the form of a matrix, where each row represents a zone and each column
represents a location. A set of zones can be generated using the
\code{\link{space_coord_to_zones}} function, and conversion between the list- and
matrix-representation of the zones is accomplished through the functions
\code{build_key_matrix} and \code{build_zones}.
}
\examples{
space_coord <- co_county_fips_2019
zones <- space_coord_to_zones(space_coord, max_k = 5)
key_matrix <- build_key_matrix(zones)

zones <- build_zones(key_matrix)
}
\seealso{
\code{\link{space_coord_to_zones}},
\code{\link[scanstatistics]{knn_zones}}, \code{\link[scanstatistics]{dist_to_knn}}
}
